+++ src/fs/axl-fs.c
@@ -87,7 +87,7 @@ axl_file_get_contents(const char *path, void **buf, size_t *len)
         (const unsigned short *)wide_path, AXL_FILE_MODE_READ, 0, &handle);
     axl_free(wide_path);
     if (rc != AXL_OK) {
-        axl_warning("open failed: %s", path);
+        axl_debug("open failed: %s", path);
         return AXL_ERR;
     }
 
@@ -168,7 +168,7 @@ axl_file_set_contents(const char *path, const void *buf, size_t len)
         0, &handle);
     axl_free(wide_path);
     if (rc != AXL_OK) {
-        axl_warning("open failed: %s", path);
+        axl_debug("open failed: %s", path);
         return AXL_ERR;
     }